+++ examples/demo/src/routes/HomePage.tsx
@@ -3,6 +3,7 @@ import {
   Recommend,
   TrendingFacets,
   TrendingItems,
+  RecommendedForYou,
 } from '@algolia/recommend-react';
 import { HorizontalSlider } from '@algolia/ui-components-horizontal-slider-react';
 import React from 'react';
@@ -10,7 +11,7 @@ import { useNavigate } from 'react-router-dom';
 
 import { Hit, Facet } from '../components';
 import { apiKey, appId, indexName } from '../config';
-import { FacetHit, ProductHit } from '../types';
+import { ProductHit } from '../types';
 
 import { useApplicationContext } from './Root';
 
@@ -24,7 +25,7 @@ export const HomePage: React.FC = () => {
 
   return (
     <Recommend recommendClient={recommendClient}>
-      <TrendingFacets<FacetHit>
+      <TrendingFacets
         indexName={indexName}
         facetName="brand"
         itemComponent={({ item }) => (
@@ -68,6 +69,24 @@ export const HomePage: React.FC = () => {
             selectedFacetValue ? `in ${selectedFacetValue.facetValue}` : ''
           }`,
         }}
+      />
+      <RecommendedForYou<ProductHit>
+        indexName={indexName}
+        maxRecommendations={15}
+        queryParameters={{
+          userToken: 'user-token-1',
+        }}
+        itemComponent={({ item }) => (
+          <Hit
+            hit={item}
+            insights={insights}
+            onSelect={(item) => {
+              setSelectedProduct(item);
+              navigate(`/product/${item.objectID}`);
+            }}
+          />
+        )}
+        fallbackComponent={() => <div>No recommendations</div>}
       />
     </Recommend>
   );