+++ examples/js-demo/app.tsx
@@ -6,12 +6,15 @@ import {
   frequentlyBoughtTogether,
   relatedProducts,
   lookingSimilar,
+  trendingFacets,
+  recommendedForYou,
 } from '@algolia/recommend-js';
 import { horizontalSlider } from '@algolia/ui-components-horizontal-slider-js';
 import algoliasearch from 'algoliasearch';
 import { h, render } from 'preact';
 import insights from 'search-insights';
 
+import { Facet } from './Facet';
 import { RelatedItem } from './RelatedItem';
 import { ProductHit, ReferenceItemProps } from './types';
 
@@ -95,6 +98,46 @@ autocomplete<ProductHit>({
       },
     ];
   },
+});
+
+trendingFacets({
+  container: '#trendingFacets',
+  indexName,
+  recommendClient,
+  facetName: 'brand',
+  maxRecommendations: 5,
+  itemComponent({ item }) {
+    return <Facet hit={item} />;
+  },
+  view: (props) => {
+    return (
+      <div className="Facets-Wrapper">
+        {props[0].items.map((item) => {
+          return <Facet key={item.facetValue} hit={item} />;
+        })}
+      </div>
+    );
+  },
+});
+
+recommendedForYou<ProductHit>({
+  container: '#recommendedForYou',
+  recommendClient,
+  indexName,
+  maxRecommendations: 15,
+  queryParameters: {
+    userToken: 'user-token-1',
+  },
+  itemComponent({ item }) {
+    return (
+      <RelatedItem
+        item={item}
+        insights={insights}
+        onSelect={updateReferenceItem}
+      />
+    );
+  },
+  view: (...props) => horizontalSlider(...props) || <div>Loading</div>,
 });
 
 function ReferenceItem({ item }: ReferenceItemProps) {
@@ -232,7 +275,6 @@ function renderRecommendations(selectedProduct: ProductHit) {
       );
     },
   });
-
   relatedProducts<ProductHit>({
     container: '#relatedProducts',
     recommendClient,