+++ packages/recommend-core/src/__tests__/getBatchRecommendations.test.ts
@@ -32,7 +32,10 @@ describe('getBatchRecommendations', () => {
 
     expect(recommendClient.getRecommendations).toHaveBeenCalledTimes(1);
     expect(Object.keys(result)[0]).toEqual('{"key":"key-1"}');
-    expect(Object.values(result)[0]).toEqual({ recommendations: [hit] });
+    expect(Object.values(result)[0]).toEqual({
+      recommendations: [hit],
+      trendingFacets: [],
+    });
   });
 
   it('should transform items', async () => {