+++ packages/recommend-core/src/getBatchRecommendations.ts
@@ -2,9 +2,10 @@ import {
   RecommendClient,
   RecommendationsQuery,
   TrendingQuery,
+  RecommendedForYouQuery,
 } from '@algolia/recommend';
 
-import { ProductRecord, TrendingFacet } from './types';
+import { TrendingFacetHit, ProductRecord } from './types';
 import { mapByScoreToRecommendations, mapToRecommendations } from './utils';
 import { version } from './version';
 
@@ -13,7 +14,11 @@ export type BatchKeyPair = {
   value: number;
 };
 
-export type BatchQuery<TObject> = (RecommendationsQuery | TrendingQuery) & {
+export type BatchQuery<TObject> = (
+  | RecommendationsQuery
+  | TrendingQuery
+  | RecommendedForYouQuery
+) & {
   transformItems?: (
     items: Array<ProductRecord<TObject>>
   ) => Array<ProductRecord<TObject>>;
@@ -27,6 +32,7 @@ export type GetBatchRecommendations<TObject> = {
 
 export type BatchRecommendations<TObject> = {
   recommendations: Array<ProductRecord<TObject>>;
+  trendingFacets: TrendingFacetHit[];
 };
 
 export async function getBatchRecommendations<TObject>({
@@ -55,10 +61,14 @@ export async function getBatchRecommendations<TObject>({
     const splitResult = response?.results?.slice(prevChunks, allChunks);
     prevChunks += keyPair.value;
 
-    let recommendations: Array<ProductRecord<ProductRecord<TObject>>>;
-
-    if (model === 'trending-facets' || model === 'trending-items') {
-      recommendations = mapByScoreToRecommendations<TrendingFacet<TObject>>({
+    let recommendations: Array<ProductRecord<ProductRecord<TObject>>> = [];
+    let trendingFacets: TrendingFacetHit[] = [];
+    if (model === 'trending-facets') {
+      trendingFacets = splitResult
+        .map((res) => (res.hits as unknown) as TrendingFacetHit[])
+        .flat();
+    } else if (model === 'trending-items') {
+      recommendations = mapByScoreToRecommendations<ProductRecord<TObject>>({
         maxRecommendations,
         hits: splitResult.map((res) => res.hits).flat(),
       });
@@ -70,7 +80,7 @@ export async function getBatchRecommendations<TObject>({
       });
     }
     recommendations = transformItems(recommendations);
-    results[keyPair.key] = { recommendations };
+    results[keyPair.key] = { recommendations, trendingFacets };
   });
 
   return results;