+++ packages/recommend-vdom/src/types/FacetsViewProps.ts
@@ -1,16 +1,15 @@
-import { FacetEntry } from '@algolia/recommend-core';
+import { TrendingFacetHit } from '@algolia/recommend-core';
 
 import { Renderer } from './Renderer';
 
 export type FacetsViewProps<
-  TItem extends FacetEntry,
   TTranslations extends Record<string, string>,
   TClassNames extends Record<string, string>
 > = {
   classNames: TClassNames;
   itemComponent<TComponentProps extends Record<string, unknown> = {}>(
-    props: { item: TItem } & Renderer & TComponentProps
+    props: { item: TrendingFacetHit } & Renderer & TComponentProps
   ): JSX.Element;
-  items: TItem[];
+  items: TrendingFacetHit[];
   translations: TTranslations;
 };